backend/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.9.0
asgiref==3.9.2
sqlparse==0.5.3
tzdata==2025.2
